+++ poem/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.42] 2022-09-11
+
+- Fixed `StaticFileEndpoint` returning an incorrect `Content-Length` header when a `Range` header is in the request.
+- Fixed `Compression` middleware returning incorrect `Content-Length` header.
+
 # [1.3.41] 2022-08-16
 
 - Use the real IP as the `remote_addr` in the logs of the Tracing middleware. [#370](https://github.com/poem-web/poem/issues/370)